refactor: :truck: added classes folder
