+++ src/classes/Ebook.ts
@@ -1,7 +1,7 @@
-import { getEBook } from "./methods/book/ebook/getEBook";
-import { getEBookmark } from "./methods/book/ebook/getEbookmark";
-import { setEBookmark } from "./methods/book/ebook/setEBookmark";
-import type { JWT, SingleSignToken } from "./types/types";
+import { getEBook } from "../methods/book/ebook/getEBook";
+import { getEBookmark } from "../methods/book/ebook/getEbookmark";
+import { setEBookmark } from "../methods/book/ebook/setEBookmark";
+import type { JWT, SingleSignToken } from "../types/types";
 
 export class Ebook {
 	constructor(
